Silence errors about hidden functions in nosqlprotocol
The compilation of the nosqlprotocol on Ubuntu Noble fails as it detects
that the virtual functions from OpMsgCommand are hidden by the ones in
SaslStart and SaslContinue. This hiding is not a practical problem but
it does cause the build to fail.

Disabling the warnings would also fix it but doing that would be more
code.