MXS-5387: Fix error in partial result handling
If transaction replay is enabled, partially executed statements where
results were already delivered to the client are attempted again in the
hopes that the interrupted query can be re-executed and spliced into the
already returned part. However, this currently requires that a
transaction is open and if done without a transaction, m_current_query
will be empty as it was freed in manage_transactions(). The code did not
check for the transaction state before continuing on to retry the query.

The two cases, multi-statments and partial results, are now in their own
branches which makes the code easier to comprehend.