MXS-5387: Fix partially delivered ignorable errors
If a result that has been partially delivered which also ends in an
error that would otherwise qualify for automatic retrying (e.g. WSREP
errors) is received, the m_current_query would be empty at the point at
which the retrying would be started. For writes, this would be detected
and internally handled in a way that would not cause a crash. For reads,
it would not check whether m_current_query was empty or not. This would
result in an empty GWBUF being used in code that assumed non-empty
GWBUFs.

Added test case that reprocudes the problem where WSREP errors caused an
unwanted retry to take place.