MXS-5394: Always return null for empty passwords
The behavior was different for empty passwords when password encryption
was enabled and disabled. When it was disabled it was returned as null
but when it was enabled it was returned as an encrypted password. This
caused the password to be "re-encrypted" every time it was displayed in
the REST-API.

The functionality of the passwords was not affected, it was simply how
they were displayed in the REST-API.