Disable detect_stack_use_after_return in aarch64 tests
Enabling detect_stack_use_after_return in two of the tests seems to slow
down the execution so much that the tests time out on aarch64. Disabling
it speeds them up to normal levels. There are two bugs that this might
be related to:

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=91101
  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=94910

The former was solved by turning on optimizations (didn't work for us)
but the latter was an actual bug fix that got fixed in GCC 10.